ft(create intro section):
create intro component
add css to intro component
create image container

[ Maintains #172320154 ]
