+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from './components/Navbar'
+import Navbar from './components/navbar/Navbar'
 import MainContent from './components/mainContent'
 import './components/styles/main.scss'
